# frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Proxy API",
    "description": "File-based YARP proxy, mock, and monitor API.",
    "version": "v1"
  },
  "paths": {
...
    "/api/proxies/{proxyId}/logs": {
      "get": {
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "proxyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusCode",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LogListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogListDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/proxies/{proxyId}/logs/{entryId}": {
      "get": {
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "proxyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LogDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/proxies/{proxyId}/stats": {
      "get": {
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "proxyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyStatsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyStatsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyStatsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/proxies/{id}": {
      "get": {
        "tags": [
          "Proxies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Proxies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProxyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProxyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProxyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Proxies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CertificateDto": {
        "type": "object",
        "properties": {
          "pfxPath": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DestinationDto": {
        "required": [
          "address"
        ],
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "clientCertificate": {
            "$ref": "#/components/schemas/CertificateDto"
          },
          "acceptAnyServerCertificate": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ListenDto": {
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "pathPrefix": {
            "type": "string",
            "nullable": true
          },
          "hosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "serverCertificate": {
            "$ref": "#/components/schemas/CertificateDto"
          }
        },
        "additionalProperties": false
      },
      "LogDetailDto": {
        "required": [
          "method",
          "mode",
          "path",
          "protocol"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "query": {
            "type": "string",
            "nullable": true
          },
          "protocol": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "mode": {
            "type": "string",
            "nullable": true
          },
          "mockName": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "requestHeaders": {
            "type": "string",
            "nullable": true
          },
          "requestBody": {
            "type": "string",
            "nullable": true
          },
          "requestBodyTruncated": {
            "type": "boolean"
          },
          "responseHeaders": {
            "type": "string",
            "nullable": true
          },
          "responseBody": {
            "type": "string",
            "nullable": true
          },
          "responseBodyTruncated": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "LogListDto": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogListItemDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LogListItemDto": {
        "required": [
          "method",
          "mode",
          "path",
          "protocol"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "query": {
            "type": "string",
            "nullable": true
          },
          "protocol": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "mode": {
            "type": "string",
            "nullable": true
          },
          "mockName": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProxyDetailDto": {
        "required": [
          "destination",
          "id",
          "listen",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "listen": {
            "$ref": "#/components/schemas/ListenDto"
          },
          "destination": {
            "$ref": "#/components/schemas/DestinationDto"
          },
          "mocksEnabled": {
            "type": "boolean"
          },
          "passthroughDelayMs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProxyStatsDto": {
        "type": "object",
        "properties": {
          "totalRequests": {
            "type": "integer",
            "format": "int64"
          },
          "mockRequests": {
            "type": "integer",
            "format": "int64"
          },
          "passthroughRequests": {
            "type": "integer",
            "format": "int64"
          },
          "averageDurationMs": {
            "type": "number",
            "format": "double"
          },
          "lastStatusCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastRequestUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpsertProxyRequest": {
        "required": [
          "destination",
          "listen",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "listen": {
            "$ref": "#/components/schemas/ListenDto"
          },
          "destination": {
            "$ref": "#/components/schemas/DestinationDto"
          },
          "mocksEnabled": {
            "type": "boolean"
          },
          "passthroughDelayMs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}